+++ 2022/11/types.hh
@@ -1,5 +1,5 @@
 #pragma once
 #include <functional>
 
-using WorryValue = int;
+using WorryValue = long long int;
 using WorryManager = std::function<WorryValue(WorryValue v)>;
